Consolidate json (de)serialization code into a single module
