+++ gufe/tests/test_custom_json.py
@@ -15,7 +15,7 @@ from numpy import testing as npt
 from openff.units import unit
 
 from gufe import tokenization
-from gufe.serialization.json.custom_codecs import (
+from gufe.serialization.json import (
     BYTES_CODEC,
     NPY_DTYPE_CODEC,
     NUMPY_CODEC,
@@ -24,8 +24,10 @@ from gufe.serialization.json.custom_codecs import (
     PATH_CODEC,
     SETTINGS_CODEC,
     UUID_CODEC,
+    JSONCodec,
+    JSONSerializerDeserializer,
+    custom_json_factory,
 )
-from gufe.serialization.json.custom_json import JSONCodec, JSONSerializerDeserializer, custom_json_factory
 from gufe.settings import models
