+++ gufe/tokenization.py
@@ -20,7 +20,7 @@ from typing import Any, BinaryIO, Dict, List, Optional, TextIO, Tuple, Union
 import networkx as nx
 from typing_extensions import Self
 
-from gufe.serialization.json.custom_codecs import (
+from gufe.serialization.json import (
     BYTES_CODEC,
     DATETIME_CODEC,
     NPY_DTYPE_CODEC,
@@ -30,8 +30,8 @@ from gufe.serialization.json.custom_codecs import (
     PATH_CODEC,
     SETTINGS_CODEC,
     UUID_CODEC,
+    JSONSerializerDeserializer,
 )
-from gufe.serialization.json.custom_json import JSONSerializerDeserializer
 from gufe.serialization.msgpack import packb, unpackb
 
 _default_json_codecs = [
